Weather_api_streaming_project/weather_function/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure.eventhub
azure.identity
azure.keyvault.secrets